Address most comments: - Move types to new mucf/Types.hh - Move MucfParticles::from_params to model.cc as a free function - Move all static data to MucfPhysics.cc - Refactor/Remove types in DTMixMucfData - Remove CELER_EXPECT(mu_minus) during MucfProcess construction
