+++ src/celeritas/mucf/data/DTMixMucfData.hh
@@ -14,6 +14,7 @@
 #include "corecel/grid/NonuniformGridData.hh"
 #include "corecel/io/Join.hh"
 #include "celeritas/Types.hh"
+#include "celeritas/mucf/Types.hh"
 #include "celeritas/phys/ParticleParams.hh"
 
 namespace celeritas
@@ -22,7 +23,7 @@ namespace celeritas
 /*!
  * ParticleIds used by the \c DTMixMucfModel .
  */
-struct MucfParticles
+struct MucfParticleIds
 {
     //! Primary
     ParticleId mu_minus;
@@ -51,46 +52,6 @@ struct MucfParticles
                && muonic_hydrogen && muonic_alpha && muonic_triton
                && muonic_helium3;
     }
-
-    //! Assign from ParticleParams (anonymous free function in the model.cc?)
-    static MucfParticles from_params(ParticleParams const& particles)
-    {
-#define MP_ADD(MEMBER)                               \
-    pids.MEMBER = particles.find(pdg::MEMBER());     \
-    if (!pids.MEMBER)                                \
-    {                                                \
-        missing.push_back({#MEMBER, pdg::MEMBER()}); \
-    }
-
-        using PairStrPdg = std::pair<std::string, PDGNumber>;
-        std::vector<PairStrPdg> missing;
-        MucfParticles pids;
-
-        MP_ADD(mu_minus);
-        MP_ADD(neutron);
-        MP_ADD(proton);
-        MP_ADD(alpha);
-        //! \todo Decide whether to implement these PDGs in PDGNumber.hh
-#if 0
-        MP_ADD(helium_3);
-        MP_ADD(muonic_hydrogen);
-        MP_ADD(muonic_alpha);
-        MP_ADD(muonic_deuteron);
-        MP_ADD(muonic_triton);
-        MP_ADD(muonic_helium3);
-#endif
-        CELER_VALIDATE(
-            missing.empty(),
-            << "missing particles required for muon-catalyzed fusion: "
-            << join(missing.begin(), missing.end(), ", ", [](PairStrPdg const& p) {
-                   return p.first + " (PDG "
-                          + std::to_string(p.second.unchecked_get()) + ')';
-               }));
-
-        return pids;
-
-#undef MP_ADD
-    }
 };
 
 //---------------------------------------------------------------------------//
@@ -103,17 +64,15 @@ struct DTMixMucfData
     template<class T>
     using Items = Collection<T, W, M>;
     template<class T>
-    using MatCycleItems = Collection<T, W, M, MuCfMatCompId>;
-    template<class T>
     using MaterialItems = Collection<T, W, M, MuCfMatCompId>;
-    using Grid = NonuniformGridRecord;
+    using GridRecord = NonuniformGridRecord;
     using CycleTimesArray = EnumArray<MucfMuonicMolecule, Array<real_type, 2>>;
 
     //! Particle IDs
-    MucfParticles particles;
+    MucfParticleIds particles;
 
     //! Muon CDF energy grid for sampling outgoing muCF muons
-    Grid muon_energy_cdf;  //! \todo Verify energy unit
+    GridRecord muon_energy_cdf;  //! \todo Verify energy unit
     Items<real_type> reals;
 
     //!@{
@@ -121,7 +80,7 @@ struct DTMixMucfData
     //! \c PhysMatId indexed by \c MuCfMatCompId
     MaterialItems<PhysMatId> matcompid_to_matid;
     //! Cycle times per material: [mat_comp_id][muonic_molecule][spin_index]
-    MatCycleItems<CycleTimesArray> cycle_times;  //!< In [s]
+    MaterialItems<CycleTimesArray> cycle_times;  //!< In [s]
     //! \todo Add mean atom spin flip times
     //! \todo Add mean atom transfer times
     //!@}
