+++ src/celeritas/mucf/interactor/DDMucfInteractor.hh
@@ -38,7 +38,7 @@ class DDMucfInteractor
     // Construct from shared and state data
     inline CELER_FUNCTION
     DDMucfInteractor(NativeCRef<DTMixMucfData> const& data,
-                     Channel const channel,
+                     Channel channel,
                      StackAllocator<Secondary>& allocate);
 
     // Sample an interaction with the given RNG
