+++ src/celeritas/mucf/interactor/TTMucfInteractor.hh
@@ -36,7 +36,7 @@ class TTMucfInteractor
     // Construct from shared and state data
     inline CELER_FUNCTION
     TTMucfInteractor(NativeCRef<DTMixMucfData> const& data,
-                     Channel const channel,
+                     Channel channel,
                      StackAllocator<Secondary>& allocate);
 
     // Sample an interaction with the given RNG
